feat: abstract away with PushNotificationPublisher
